refactor load_stream, parameters to stream type
